Added ability to validate objects
When `new Validate\Init($directory, $temporary_file, $production)` is called, classes with annotations @Validate that are inside `$directory` is loaded, and validators for each property is generated and saved inside `$temporary_file`. The object has a method `validate($object)` that is going to return an array with all the validations errors for that object, if it is empty it means that all properties are valid.